# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation POS API",
        "description": "This API documentation outlines the features and endpoints of a Laravel project designed to master RESTful API development using Sanctum. It provides comprehensive documentation for secure user authentication, token-based sessions, and role-based access control. Developed as a learning tool, this project demonstrates best practices in Laravel API design, including resourceful routing, middleware application, and data validation techniques. It serves as a practical guide for developers aiming to implement modern API solutions with Laravel and Sanctum",
        "contact": {
            "email": "pichoendoo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/members": {
            "get": {
                "tags": [
                    "Members"
                ],
                "summary": "List members",
                "description": "Get a paginated list of members",
                "operationId": "getMembers",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Members fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Member"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Members fetched successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to fetch members. Please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Members"
                ],
                "summary": "Create member",
                "description": "Create a new member",
                "operationId": "storeMember",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreMemberRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Member created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Member"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Member created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to create member. Please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/members/{member}": {
            "get": {
                "tags": [
                    "Members"
                ],
                "summary": "Get member",
                "description": "Fetch a specific member by ID",
                "operationId": "showMember",
                "parameters": [
                    {
                        "name": "member",
                        "in": "path",
                        "description": "Member ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Member fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Member"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Member fetched successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to fetch member. Please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Members"
                ],
                "summary": "Update member",
                "description": "Update an existing member",
                "operationId": "updateMember",
                "parameters": [
                    {
                        "name": "member",
                        "in": "path",
                        "description": "Member ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateMemberRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Member updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Member"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Member updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to update member. Please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Members"
                ],
                "summary": "Delete member",
                "description": "Delete a specific member by ID",
                "operationId": "deleteMember",
                "parameters": [
                    {
                        "name": "member",
                        "in": "path",
                        "description": "Member ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Member deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Member deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to delete member. Please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreMemberRequest": {
                "title": "StoreMemberRequest",
                "description": "Store member request body data",
                "required": [
                    "name",
                    "phone_no"
                ],
                "type": "object"
            },
...
            "UpdateMemberRequest": {
                "title": "Update Member Request",
                "required": [
                    "name",
                    "phone_no"
                ],
                "properties": {
                    "name": {
                        "description": "Member's name",
                        "type": "string",
                        "maxLength": 86
                    },
                    "phone_no": {
                        "description": "Member's phone number",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Member": {
                "title": "Member",
                "description": "Member model",
                "properties": {
                    "id": {
                        "description": "ID of the member",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "description": "Name of the member",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "phone_no": {
                        "description": "Phone number of the member",
                        "type": "string",
                        "example": "123-456-7890"
                    },
                    "code": {
                        "description": "Generated code for the member",
                        "type": "string",
                        "example": "MEM123"
                    },
                    "point": {
                        "description": "Points of the member",
                        "type": "integer",
                        "example": "100"
                    },
                    "uuid": {
                        "description": "UUID of the member",
                        "type": "string",
                        "example": "a1b2c3d4-e5f6-7890"
                    },
                    "created_by": {
                        "description": "ID of the user who created the member",
                        "type": "integer",
                        "example": "1"
                    },
                    "updated_by": {
                        "description": "ID of the user who updated the member",
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}